feat: sign builder submissions
