+++ bolt-sidecar/src/builder/payload_builder.rs
@@ -1,24 +1,11 @@
 #![allow(missing_docs)]
 
-use std::str::FromStr;
-
-use alloy_eips::{
-    calc_excess_blob_gas, calc_next_block_base_fee, eip1559::BaseFeeParams, eip4895::Withdrawal,
-};
+use super::compat::{to_alloy_execution_payload, to_reth_withdrawal};
+use alloy_eips::{calc_excess_blob_gas, calc_next_block_base_fee, eip1559::BaseFeeParams};
 use alloy_primitives::{Address, Bytes, B256, U256};
 use alloy_rpc_types::Block;
-use alloy_rpc_types_engine::{
-    ExecutionPayload as AlloyExecutionPayload, ExecutionPayloadV1, ExecutionPayloadV2,
-    ExecutionPayloadV3,
-};
+use alloy_rpc_types_engine::ExecutionPayload as AlloyExecutionPayload;
 use beacon_api_client::{BlockId, StateId};
-use ethereum_consensus::{
-    capella::spec,
-    crypto::bls::{PublicKey as BlsPublicKey, SecretKey as BlsSecretKey},
-    deneb::mainnet::ExecutionPayloadHeader as ConsensusExecutionPayloadHeader,
-    ssz::prelude::{ssz_rs, ByteList, ByteVector},
-    types::mainnet::ExecutionPayload as ConsensusExecutionPayload,
-};
 use hex::FromHex;
 use regex::Regex;
 use reth_primitives::{
@@ -28,7 +15,6 @@ use reth_primitives::{
 use reth_rpc_layer::{secret_to_bearer_header, JwtSecret};
 use serde_json::Value;
 
-use crate::primitives::{BuilderBid, SignedBuilderBid};
 use crate::RpcClient;
 
 #[derive(Debug, thiserror::Error)]
@@ -65,22 +51,18 @@ impl FallbackPayloadBuilder {
     pub fn new(
         fee_recipient: Address,
         jwt_hex: &str,
-        engine_rpc_url: Option<&str>,
-        execution_rpc_url: Option<&str>,
+        engine_rpc_url: &str,
+        execution_rpc_url: &str,
     ) -> Self {
         Self {
             fee_recipient,
             engine_hinter: EngineHinter {
                 client: reqwest::Client::new(),
                 jwt_hex: jwt_hex.to_string(),
-                engine_rpc_url: engine_rpc_url
-                    .unwrap_or("http://localhost:8551")
-                    .to_string(),
+                engine_rpc_url: engine_rpc_url.to_string(),
             },
             extra_data: hex::encode("Selfbuilt w Bolt").into(),
-            execution_rpc_client: RpcClient::new(
-                execution_rpc_url.unwrap_or("http://localhost:8545"),
-            ),
+            execution_rpc_client: RpcClient::new(execution_rpc_url),
         }
     }
 }
@@ -115,7 +97,7 @@ impl FallbackPayloadBuilder {
     ) -> Result<(SealedBlock, SealedHeader), PayloadBuilderError> {
         let latest_block = self.execution_rpc_client.get_block(None, true).await?;
 
-        // TODO: refactor this once ConsensusState is ready
+        // TODO: refactor this once ConsensusState (https://github.com/chainbound/bolt/issues/58) is ready
         let beacon_api_endpoint = reqwest::Url::parse("http://remotebeast:3500").unwrap();
         let beacon_api = beacon_api_client::mainnet::Client::new(beacon_api_endpoint);
 
@@ -133,22 +115,23 @@ impl FallbackPayloadBuilder {
             Some(withdrawals)
         };
 
-        let prev_randao: Value = reqwest::Client::new()
+        // NOTE: for some reason, this call fails with an ApiResult deserialization error
+        // when using the beacon_api_client crate directly, so we use reqwest temporarily.
+        // this is to be refactored.
+        let prev_randao = reqwest::Client::new()
             .get("http://remotebeast:3500/eth/v1/beacon/states/head/randao")
             .send()
             .await
             .unwrap()
-            .json()
+            .json::<Value>()
             .await
             .unwrap();
-        let prev_randao = B256::from_hex(
-            prev_randao
-                .pointer("/data/randao")
-                .unwrap()
-                .as_str()
-                .unwrap(),
-        )
-        .unwrap();
+        let prev_randao = prev_randao
+            .pointer("/data/randao")
+            .unwrap()
+            .as_str()
+            .unwrap();
+        let prev_randao = B256::from_hex(prev_randao).unwrap();
 
         let parent_beacon_block_root = beacon_api
             .get_beacon_block_root(BlockId::Head)
@@ -225,22 +208,6 @@ impl FallbackPayloadBuilder {
 
             i += 1;
         };
-
-        // NOTE: this should be fine as the beacon node cannot validate it
-        // let value = U256::from_str("1000000000000000000").expect("valid int");
-
-        // let submission = BuilderBid {
-        //     header: to_execution_payload_header(&sealed_header),
-        //     blob_kzg_commitments: List::default(),
-        //     public_key: self.public_key.clone(),
-        //     value: ctx.value,
-        // };
-
-        // let signed_submission = SignedBuilderBid {
-        //     message: submission,
-        //     // TODO: sign the message
-        //     signature: Default::default(),
-        // };
 
         Ok((sealed_block, sealed_header))
     }
@@ -367,143 +334,6 @@ fn build_header_with_hints_and_context(
     }
 }
 
-/// Compatibility: convert a sealed header into an ethereum-consensus execution payload header
-pub(crate) fn to_execution_payload_header(value: &SealedHeader) -> ConsensusExecutionPayloadHeader {
-    ConsensusExecutionPayloadHeader {
-        parent_hash: to_bytes32(value.parent_hash),
-        fee_recipient: to_bytes20(value.beneficiary),
-        state_root: to_bytes32(value.state_root),
-        receipts_root: to_bytes32(value.receipts_root),
-        logs_bloom: to_byte_vector(value.logs_bloom),
-        prev_randao: to_bytes32(value.mix_hash),
-        block_number: value.number,
-        gas_limit: value.gas_limit,
-        gas_used: value.gas_used,
-        timestamp: value.timestamp,
-        extra_data: ByteList::try_from(value.extra_data.as_ref()).unwrap(),
-        base_fee_per_gas: ssz_rs::U256::from(value.base_fee_per_gas.unwrap_or_default()),
-        block_hash: to_bytes32(value.hash()),
-        transactions_root: value.transactions_root,
-        withdrawals_root: value.withdrawals_root.unwrap_or_default(),
-        blob_gas_used: value.blob_gas_used.unwrap_or_default(),
-        excess_blob_gas: value.excess_blob_gas.unwrap_or_default(),
-    }
-}
-
-/// Compatibility: convert a sealed block into an Alloy execution payload
-pub(crate) fn to_alloy_execution_payload(
-    block: &SealedBlock,
-    block_hash: B256,
-) -> AlloyExecutionPayload {
-    println!("block hash in to_alloy_execution_payload: {:?}", block_hash);
-
-    let alloy_withdrawals = block
-        .withdrawals
-        .as_ref()
-        .map(|withdrawals| {
-            withdrawals
-                .iter()
-                .map(|w| Withdrawal {
-                    index: w.index,
-                    validator_index: w.validator_index,
-                    address: w.address,
-                    amount: w.amount,
-                })
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default();
-
-    AlloyExecutionPayload::V3(ExecutionPayloadV3 {
-        blob_gas_used: block.blob_gas_used(),
-        excess_blob_gas: block.excess_blob_gas.unwrap_or_default(),
-        payload_inner: ExecutionPayloadV2 {
-            payload_inner: ExecutionPayloadV1 {
-                base_fee_per_gas: U256::from(block.base_fee_per_gas.unwrap_or_default()),
-                block_hash,
-                block_number: block.number,
-                extra_data: block.extra_data.clone(),
-                transactions: block.raw_transactions(),
-                fee_recipient: block.header.beneficiary,
-                gas_limit: block.gas_limit,
-                gas_used: block.gas_used,
-                logs_bloom: block.logs_bloom,
-                parent_hash: block.parent_hash,
-                prev_randao: block.mix_hash,
-                receipts_root: block.receipts_root,
-                state_root: block.state_root,
-                timestamp: block.timestamp,
-            },
-            withdrawals: alloy_withdrawals,
-        },
-    })
-}
-
-/// Compatibility: convert a sealed block into an ethereum-consensus execution payload
-pub(crate) fn to_consensus_execution_payload(value: &SealedBlock) -> ConsensusExecutionPayload {
-    let hash = value.hash();
-    let header = &value.header;
-    let transactions = &value.body;
-    let withdrawals = &value.withdrawals;
-    let transactions = transactions
-        .iter()
-        .map(|t| spec::Transaction::try_from(t.envelope_encoded().as_ref()).unwrap())
-        .collect::<Vec<_>>();
-    let withdrawals = withdrawals
-        .as_ref()
-        .unwrap_or(&Withdrawals::default())
-        .iter()
-        .map(|w| spec::Withdrawal {
-            index: w.index as usize,
-            validator_index: w.validator_index as usize,
-            address: to_bytes20(w.address),
-            amount: w.amount,
-        })
-        .collect::<Vec<_>>();
-
-    let payload = spec::ExecutionPayload {
-        parent_hash: to_bytes32(header.parent_hash),
-        fee_recipient: to_bytes20(header.beneficiary),
-        state_root: to_bytes32(header.state_root),
-        receipts_root: to_bytes32(header.receipts_root),
-        logs_bloom: to_byte_vector(header.logs_bloom),
-        prev_randao: to_bytes32(header.mix_hash),
-        block_number: header.number,
-        gas_limit: header.gas_limit,
-        gas_used: header.gas_used,
-        timestamp: header.timestamp,
-        extra_data: ByteList::try_from(header.extra_data.as_ref()).unwrap(),
-        base_fee_per_gas: ssz_rs::U256::from(header.base_fee_per_gas.unwrap_or_default()),
-        block_hash: to_bytes32(hash),
-        transactions: TryFrom::try_from(transactions).unwrap(),
-        withdrawals: TryFrom::try_from(withdrawals).unwrap(),
-    };
-    ConsensusExecutionPayload::Capella(payload)
-}
-
-/// Compatibility: convert a withdrawal from ethereum-consensus to a Reth withdrawal
-fn to_reth_withdrawal(
-    value: ethereum_consensus::capella::Withdrawal,
-) -> reth_primitives::Withdrawal {
-    reth_primitives::Withdrawal {
-        index: value.index as u64,
-        validator_index: value.validator_index as u64,
-        address: Address::from_slice(value.address.as_ref()),
-        amount: value.amount,
-    }
-}
-
-fn to_bytes32(value: B256) -> spec::Bytes32 {
-    spec::Bytes32::try_from(value.as_ref()).unwrap()
-}
-
-fn to_bytes20(value: Address) -> spec::ExecutionAddress {
-    spec::ExecutionAddress::try_from(value.as_ref()).unwrap()
-}
-
-fn to_byte_vector(value: Bloom) -> ByteVector<256> {
-    ByteVector::<256>::try_from(value.as_ref()).unwrap()
-}
-
 #[cfg(test)]
 mod tests {
     use alloy_eips::eip2718::Encodable2718;
@@ -526,8 +356,7 @@ mod tests {
         let execution = "http://remotebeast:8545";
         let engine = "http://remotebeast:8551";
 
-        let builder =
-            FallbackPayloadBuilder::new(Address::default(), &jwt, Some(engine), Some(execution));
+        let builder = FallbackPayloadBuilder::new(Address::default(), &jwt, engine, execution);
 
         let sk = SigningKey::from_slice(hex::decode(raw_sk)?.as_slice())?;
         let signer = PrivateKeySigner::from_signing_key(sk.clone());
